Refactor navigation setup into an Eleventy plugin
